+++ src/main/java/racingcar/domain/RandomNumberGenerator.java
@@ -1,6 +1,8 @@
-package racingcar;
+package racingcar.domain;
 
 import java.util.Random;
+
+import racingcar.domain.NumberGenerator;
 
 public class RandomNumberGenerator implements NumberGenerator {
